using LocalDate in DB

+++ src/main/java/com/upgrade/challenge/exception/BookingNotFoundException.java
@@ -6,11 +6,11 @@ public class BookingNotFoundException extends RuntimeException {
 
 	private static final String MESSAGE = "Booking not found for ID: %d";
 
-	public BookingNotFoundException(Integer id, Throwable cause) {
+	public BookingNotFoundException(Long id, Throwable cause) {
 		super(String.format(MESSAGE, id), cause);
 	}
 
-	public BookingNotFoundException(Integer id) {
+	public BookingNotFoundException(Long id) {
 		super(String.format(MESSAGE, id));
 	}
