+++ src/main/java/com/upgrade/challenge/repository/BookingRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.upgrade.challenge.model.dto.Booking;
 
 @Repository
-public interface BookingRepository extends CrudRepository<Booking, Integer> {
+public interface BookingRepository extends CrudRepository<Booking, Long> {
 
 }
